safety/audit_and_rollback.py
# Import required modules
import logging
from safety.function_execution import execute_function  # Replace 'some_module' with the actual module where execute_function is defined
from utils.utils import print_tracer  # Import print_tracer for debugging and context

# Optional: Function to take a system snapshot
def take_system_snapshot():
    """
    This function takes a snapshot of the system state.
    """
    print_tracer("audit_and_rollback", "take_system_snapshot", "Start")
    # Your snapshot logic here
    print_tracer("audit_and_rollback", "take_system_snapshot", "End")
    return "Snapshot Taken"  # Replace with actual snapshot data

# Function to execute a function with audit
def execute_function_with_audit(func_call):
    """
    This function executes a given function call and performs auditing.
    """
    print_tracer("audit_and_rollback", "execute_function_with_audit", "Start")

    # Optional: Take a snapshot of the system state
    snapshot = take_system_snapshot()

    # Execute the function and capture the outcome
    outcome = execute_function(func_call)

    # Log the executed function and outcome
    logging.info(f"Executed {func_call}, Outcome: {outcome}")

    print_tracer("audit_and_rollback", "execute_function_with_audit", "End")

    return outcome  # You can return the outcome or any other data you need
