tests/saftey/test_audit_and_rollback.py
import unittest
import logging
from unittest import mock
from safety.audit_and_rollback import execute_function_with_audit, take_system_snapshot

class TestAuditAndRollback(unittest.TestCase):

    @mock.patch('safety.audit_and_rollback.execute_function')
    @mock.patch('safety.audit_and_rollback.take_system_snapshot')
    def test_execute_function_with_audit(self, mock_snapshot, mock_execute):
        mock_snapshot.return_value = "Snapshot Taken"
        mock_execute.return_value = "Function Executed"

        result = execute_function_with_audit("TestFunction")

        mock_snapshot.assert_called_once()
        mock_execute.assert_called_once_with("TestFunction")
        self.assertEqual(result, "Function Executed")
        # Add any more assertions you need

    def test_take_system_snapshot(self):
        result = take_system_snapshot()
        self.assertEqual(result, "Snapshot Taken")
        # Add any more assertions you need

if __name__ == '__main__':
    logging.basicConfig(level=logging.INFO)  # This will allow you to see your logging info statements
    unittest.main()
